general utility code moved to NetscapeBookmarksFile class.
